Last major changes
- removed components hiding with client-side JavaScript in favor of just CSS media queries.

- changed all pages using an api route to fetch the current user to use a function that simply extrracts a cookie from any request.

- also brought out most of said functions from each individual component, and instead put it in the layout files to feed them as props to the components.

- completed admin page, which now functions a lot like the normal search page with some extra features for management.

 - added displaying of errors in most forms like login and account creation.

 - added API routes for admin use along with functions to check for admin authority when using said functions.

 - Settings page finished users can now update their account information or delete their account.

 - also added a fake light mode button that logs you out because screw light mode and anyone who wants to use it.

- Other unlisted changes.

TODO:
Thoroughly looks through all doe and improve/clean it up and fix any oversights that might have been made

+++ src/pages/api/comments/deletecomment.ts
@@ -1,33 +1,32 @@
 import type { APIRoute } from "astro";
 import { importJWK } from "jose";
+import { getSecret } from "~/server/misc";
 import { getCommentAuthorId, deleteComment } from "~/server/posts";
 import { verifyToken } from "~/server/users";
 
 export const DELETE: APIRoute = async ({ request, cookies }) => {
-    const body = await request.json();
-    const commentId = String(body.id);
-    const authorId = String(await getCommentAuthorId(commentId));
+  const body = await request.json();
+  const commentId = String(body.id);
+  const authorId = String(await getCommentAuthorId(commentId));
 
-    const authToken = cookies.get("authToken")?.value;
+  const authToken = cookies.get("authToken")?.value;
   if (!authToken)
     return new Response(JSON.stringify("not work"), { status: 401 });
 
-  const secretKey = await importJWK({
-    kty: "oct",
-    k: import.meta.env.JWT_SECRET,
-    alg: "HS256",
-  });
+  const secretKey = await getSecret();
 
   const { payload } = await verifyToken(authToken, secretKey);
   if (!payload) return new Response(JSON.stringify("nope"), { status: 401 });
 
-  const userId = payload.sub?.replace(/['"]+/g, "");
+  const userId = payload.sub;
+  if (!userId)
+    return new Response(JSON.stringify("Token error"), { status: 401 });
+
   if (userId == authorId) {
     await deleteComment(commentId);
     return new Response(JSON.stringify(body), {
-        status: 200,
-      });
-      
+      status: 200,
+    });
   }
-  return new Response(JSON.stringify(body))
-}
+  return new Response(JSON.stringify(body));
+};
